Failure of Post parameters
I didn't really have issues until I got to the post create. This is erroring due to 'param is missing or the value is empty : user'. This is a little confusing since I am not sure what it needs to pass in here. From what I can tell it only needs the name, email, and phone_number. I am not sure if I should just pass in a User.new() like what I see online bu I am not sure if that is what I need to do here.